+++ tests/mini.c
@@ -47,7 +47,7 @@ size_t to_client_len;
 static ssize_t
 client_pull (gnutls_transport_ptr_t tr, void *data, size_t len)
 {
-  success ("client_pull len %d has %d\n", len, to_client_len);
+  success ("client_pull len %d has %d\n", (int)len, (int)to_client_len);
 
   if (to_client_len < len)
     {
@@ -69,7 +69,7 @@ client_push (gnutls_transport_ptr_t tr, const void *data, size_t len)
   size_t newlen = to_server_len + len;
   char *tmp;
 
-  success ("client_push len %d has %d\n", len, to_server_len);
+  success ("client_push len %d has %d\n", (int)len, (int)to_server_len);
   hexprint (data, len);
 
   tmp = realloc (to_server, newlen);
@@ -89,7 +89,7 @@ client_push (gnutls_transport_ptr_t tr, const void *data, size_t len)
 static ssize_t
 server_pull (gnutls_transport_ptr_t tr, void *data, size_t len)
 {
-  success ("server_pull len %d has %d\n", len, to_server_len);
+  success ("server_pull len %d has %d\n", (int)len, (int)to_server_len);
 
   if (to_server_len < len)
     {
@@ -111,7 +111,7 @@ server_push (gnutls_transport_ptr_t tr, const void *data, size_t len)
   size_t newlen = to_client_len + len;
   char *tmp;
 
-  success ("server_push len %d has %d\n", len, to_client_len);
+  success ("server_push len %d has %d\n", (int)len, (int)to_client_len);
 
   hexprint (data, len);
 
@@ -200,7 +200,7 @@ doit (void)
   success ("Handshake established\n");
 
   ns = gnutls_record_send (client, MSG, strlen (MSG));
-  success ("client: sent %d\n", ns);
+  success ("client: sent %d\n", (int)ns);
 
   ret = gnutls_record_recv (server, buffer, MAX_BUF);
   if (ret == 0)
@@ -216,7 +216,7 @@ doit (void)
     }
 
   ns = gnutls_record_send (server, MSG, strlen (MSG));
-  success ("server: sent %d\n", ns);
+  success ("server: sent %d\n", (int)ns);
 
   ret = gnutls_record_recv (client, buffer, MAX_BUF);
   if (ret == 0)
