+++ tests/x509sign-verify.c
@@ -147,7 +147,7 @@ doit (void)
 
   for (i = 0; i < sizeof (key_dat) / sizeof (key_dat[0]); i++)
     {
-      success ("loop %d\n", i);
+      success ("loop %d\n", (int)i);
 
       ret = gnutls_x509_privkey_init (&key);
       if (ret < 0)
